requirements.txt
discord.py==2.3.2
requests==2.31.0
beautifulsoup4==4.9.3
python-dotenv==1.0.0
